pyproject.toml
[tool.poetry]
name = "udacity-capstone"
version = "0.1.0"
description = ""
authors = ["Victor Costa <victor.costa@loft.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
mypy = "^0.782"
pydantic = "^1.5.1"
apache-airflow = "^1.10.10"
pyspark = "^3.0.0"
matplotlib = "^3.2.2"
black = "^19.10b0"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
